Extract classes in to separate files
